[164829313] Reconfigures views to interact with Redux
